requirements.txt
certifi>=2019.11.28
chardet>=3.0.4
future>=0.18.2
huepy>=1.2.1
idna>=2.8
pysocks>=1.7.1
pytz>=2019.3
requests>=2.22.0
requests-toolbelt>=0.9.1
responses>=0.10.9
schedule>=0.6.0
six>=1.14.0
tqdm>=4.41.1
urllib3>=1.25.7
mock>=3.0.5
moviepy>=1.0.1
Pillow>=6.2.2
pytest>=4.6.9
rsa>=4.0
pycryptodome>=3.9.7
pycryptodomex>=3.9.7
python-dotenv
bs4
imap_tools
html5lib
zstd